refactor(codemod): pin the --json field contract on stdout instead of on a wrapper
formatJson wrapped one JSON.stringify call in a module and a 62-line test. Three
of those four tests asserted JSON.stringify's own behaviour; the fourth pinned the
field set one layer above where a consumer reads it, so it could not see the CLI
picking the wrong formatter or anything else writing to stdout.

The key set is now asserted on the real process output. formatHuman keeps its unit
tests — it has five branches worth localizing.
